일정 삭제

+++ src/main/webapp/schedule/js/details.js
@@ -56,7 +56,7 @@ document.addEventListener('DOMContentLoaded', () => {
 			let deleteBtn = '<td></td>'
 			
 			if (userCanDelete) {
-				deleteBtn = `<td><a href="${CONTEXT_PATH}/DeleteDetails?detail_id=${item.id}">삭제</a></td>`;
+				deleteBtn = `<td><a href="${CONTEXT_PATH}/DeleteDetails?detail_id=${item.id}" onclick="return confirm('정말 삭제하시겠습니까?');">삭제</a></td>`;
 			}
             
             row.innerHTML = `
